Split save_restore.rs (635 lines) into save_restore/ subdirectory
Split crates/save/src/save_restore.rs into focused modules under save_restore/:
- mod.rs: thin facade with mod declarations and pub use re-exports
- infrastructure.rs: road segments, stormwater, storm drainage, flood, CSO/sewer
- weather.rs: weather, climate, degree days, wind damage, drought, heat wave, cold snap, fog, snow
- economy.rs: policies, budget, loans, unlock state, lifecycle, virtual population, agriculture
- water.rs: water sources, reservoir, treatment, groundwater, wastewater, conservation
- waste.rs: recycling, composting, hazardous waste, landfill gas, landfill capacity, UHI grid

All public items re-exported from mod.rs — no external import changes needed.
